criu: wrap compatible code with CONFIG_COMPAT
As we will not support *native* 32-bit i386 build, I dropped some
ifdefs, introducing instead of ifdef(CONFIG_X86_64) --
ifdef(CONFIG_COMPAT).

CRIU should build now without ia32 Glibc and gcc-multilib support.
Maybe I should add some warning that the build is made without
compatible task support (or criu check should print that),
will do as compat patches hit master.

Fixes: #181

Cc: Cyrill Gorcunov <gorcunov@openvz.org>
Cc: Andrew Vagin <avagin@virtuozzo.com>
Acked-by: Andrew Vagin <avagin@virtuozzo.com>

+++ criu/page-pipe.c
@@ -501,7 +501,7 @@ void debug_show_page_pipe(struct page_pipe *pp)
 			if (pp->flags & PP_COMPAT) {
 				iov_c = (void *)ppb->iov;
 				pr_debug("\t\t%x %lu\n", iov_c[i].iov_base,
-						iov_c[i].iov_len / PAGE_SIZE);
+				   (unsigned long)iov_c[i].iov_len / PAGE_SIZE);
 			} else {
 				iov = &ppb->iov[i];
 				pr_debug("\t\t%p %lu\n", iov->iov_base,
@@ -515,7 +515,7 @@ void debug_show_page_pipe(struct page_pipe *pp)
 		if (pp->flags & PP_COMPAT) {
 			iov_c = (void *)pp->holes;
 			pr_debug("\t%x %lu\n", iov_c[i].iov_base,
-					iov_c[i].iov_len / PAGE_SIZE);
+				(unsigned long)iov_c[i].iov_len / PAGE_SIZE);
 		} else {
 			iov = &pp->holes[i];
 			pr_debug("\t%p %lu\n", iov->iov_base,
